requirements.txt
uwsgi>=2.0.18
psycopg2>=2.7,<3.0
werkzeug==0.16.1
# Install swagger-ui before connexion.
connexion[swagger-ui]
connexion == 2.6.0
Flask-SQLAlchemy
# https://realpython.com/flask-connexion-rest-api-part-2/ start
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
# https://realpython.com/flask-connexion-rest-api-part-2/ end
GeoAlchemy2>=0.7.0
python_dateutil == 2.6.0